source_system/requirements.txt
psycopg2-binary==2.9.9
pandas==2.2.1
Faker==24.1.0
streamlit==1.55.0
pydeck==0.9.1